170822 17:24 폴더 생성 완료. todo 비트맵을 이미지파일로 저장 및 관리

+++ app/src/main/java/hanbat/encho/com/notificationcollactor/NotificationListener.java
@@ -4,10 +4,12 @@ import android.app.Notification;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.os.Environment;
 import android.service.notification.NotificationListenerService;
 import android.service.notification.StatusBarNotification;
 import android.util.Log;
@@ -68,6 +70,8 @@ public class NotificationListener extends NotificationListenerService {
             String packageName = sbn.getPackageName();
 
             CharSequence appName = pm.getApplicationLabel(app);
+            Drawable d = pm.getApplicationIcon(app);
+            saveBitmapToFile(convertDrawableToBitmap(d), packageName);
 
             NotificationObject obj = new NotificationObject(title, smallIcon, largeIcon, text, subText, postTime, packageName, appName);
 
@@ -87,18 +91,34 @@ public class NotificationListener extends NotificationListenerService {
         Log.d("hanlog enter", "");
     }
 
+    private Bitmap convertDrawableToBitmap(Drawable d) {
+        if (d instanceof BitmapDrawable)
+            return ((BitmapDrawable) d).getBitmap();
+
+        Bitmap b = Bitmap.createBitmap(d.getIntrinsicWidth(), d.getIntrinsicHeight(), Bitmap.Config.ARGB_8888);
+        Canvas canvas = new Canvas(b);
+        d.setBounds(0, 0, canvas.getWidth(), canvas.getHeight());
+        d.draw(canvas);
+
+        return b;
+    }
 
     private void saveBitmapToFile(Bitmap bm, String filePath) {
-        File f = new File(filePath);
-        OutputStream stream = null;
+        File f = new File(getExternalCacheDir(), "IconImage");
+        if (!f.mkdirs())
+            Log.d("hanlog mkdir check", "FAILURE");
 
-        try {
-            f.createNewFile();
-            stream = new FileOutputStream(f);
+        File iconImage = new File(f, filePath + ".png");
+        if (!iconImage.exists()) {
+            OutputStream stream;
+            try {
+                f.createNewFile();
+                stream = new FileOutputStream(f);
 
-            bm.compress(Bitmap.CompressFormat.PNG, 100, stream);
-        } catch (IOException e) {
-            e.printStackTrace();
+                bm.compress(Bitmap.CompressFormat.PNG, 100, stream);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
